requirements.txt
numpy
pytorch
torchvision
matplotlib
mayavi
vtk
scikit-learn
dash
dash-bootstrap-components
plotly